fix bytes size
several problems here:
1) inconsistent API post/get value due to transformation. User created buffer with value 1000, and gets it back as "1 kb"
   this is problem to tf provider
   however in db we save calculated bytes size for the buffer, not original user input, and do not plan to change that

   solution: so we at least add documentation to api/terraform for user to know expected format to avoid tf diff

2) bytesSize in model is of type uint32:
  this corrupted data for input greater than 4.5 gb -> having buffer greater than that was not possible (this is ok for buffer), but such an input produced unexpected value in db

   solution:
   - validate that buffer size is not greater than 4gb
   - for other occurrences change bytesSize underlying type from uint32 to uint64 to allow greater values (i.e. for yt runtime ram)
   - add tests for api value -> bytesSize -> db value -> api value conversion
commit_hash:debcb1d30bc9506bd79053b0dc7486baeab6be5d

+++ pkg/abstract/model/endpoint_common.go
@@ -15,7 +15,7 @@ const (
 )
 
 type (
-	BytesSize    uint32
+	BytesSize    uint64
 	SecretString string
 )
